#.Modify other.
git-tfs-id: [http://serv-tfs:8080/tfs/SoaringLoong]$/SloongLibrary;C73

+++ Universal/Universal/LogSystem.cpp
@@ -217,7 +217,7 @@ DWORD _stdcall CLogSystem::Write(LPCTSTR szMessage)
 	Lock();
 	if (IsOpen())
 	{
-		WriteFile(m_hFileHandle, szMessage, _tcslen(szMessage)*sizeof(TCHAR), &dwWriteLength, NULL);
+		WriteFile(m_hFileHandle, szMessage, (DWORD)_tcslen(szMessage)*sizeof(TCHAR), &dwWriteLength, NULL);
 		FlushFileBuffers(m_hFileHandle);
 	}
